requirements.txt
flask==2.2.5
gunicorn==20.1.0
openai==0.28.0
Pillow==9.5.0
requests==2.28.2
replicate==0.15.4